added product inventory availability check
