+++ product-service/src/main/java/com/aithmetic/product/repository/ProductRepository.java
@@ -7,4 +7,6 @@ import java.util.List;
 
 public interface ProductRepository extends MongoRepository<Product,String> {
     List<Product> findByProductId(String productId);
+
+
 }
